Simplify peristent key handling
Persist only the private key and derive the public key on startup. The
public key needs to be derived from the persisted private key on startup
anyway to verify it. Persisting it means there are two files and
complicates the code.

Also some other refactoring to introduce ECKeyPair base class.